refactor: TypeDouble expressions may evaluate to BigRational, add proper casts
This allows storage of BigRational values in StateValues / Values vectors, e.g.,
to store constants that have been evaluated exactly.

TypeDouble.castValueTo now returns a Number instead of a Double, requiring the use
of the doubleValue() method in several places where the value is evaluated using
double arithmetic.

+++ prism/src/explicit/StateValues.java
@@ -139,7 +139,7 @@ public class StateValues implements StateVector
 				valuesI[i] = initI;
 		} else if (type instanceof TypeDouble) {
 			valuesD = new double[size];
-			Double objD = ((TypeDouble) type).castValueTo(init);
+			Double objD = ((TypeDouble) type).castValueTo(init).doubleValue();
 			double initD = objD.doubleValue();
 			for (i = 0; i < size; i++)
 				valuesD[i] = initD;
